@@ -42,7 +42,6 @@ Update noise values in SDarn fields when these can be calculated.
 """
 import logging
 import numpy as np
-import deepdish as dd
 
 from datetime import datetime
 from typing import Union
@@ -818,23 +817,27 @@ class BorealisConvert(BorealisRead):
         (record_key, record_dict) = borealis_rawacf_record
 
         shaped_data = {}
+        data_dimensions = record_dict.get('correlation_dimensions', None)     # Until Borealis v0.7
+        if data_dimensions is None:
+            data_dimensions = record_dict.get('data_dimensions')        # Borealis v0.7 onwards
+
         # correlation_descriptors are num_beams, num_ranges, num_lags
         # scale by the scale squared to make up for the multiply
         # in correlation (integer max squared)
         shaped_data['main_acfs'] = record_dict['main_acfs'].reshape(
-            record_dict['correlation_dimensions']).astype(
+            data_dimensions).astype(
             np.complex64) *\
             ((np.iinfo(np.int16).max**2 * scaling_factor) /
              (record_dict['data_normalization_factor']**2))
 
         if 'intf_acfs' in record_dict.keys():
             shaped_data['intf_acfs'] = record_dict['intf_acfs'].reshape(
-                record_dict['correlation_dimensions']).astype(np.complex64) *\
+                data_dimensions).astype(np.complex64) *\
                 ((np.iinfo(np.int16).max**2 * scaling_factor) /
                  (record_dict['data_normalization_factor']**2))
         if 'xcfs' in record_dict.keys():
             shaped_data['xcfs'] = record_dict['xcfs'].reshape(
-                record_dict['correlation_dimensions']).astype(np.complex64) *\
+                data_dimensions).astype(np.complex64) *\
                 ((np.iinfo(np.int16).max**2 * scaling_factor) /
                  (record_dict['data_normalization_factor']**2))
 
@@ -882,8 +885,8 @@ class BorealisConvert(BorealisRead):
                 # in Borealis file because Borealis keeps alternate
                 # lag0
                 new_data = int_data.reshape(
-                    record_dict['correlation_dimensions'][1],
-                    record_dict['correlation_dimensions'][2]-1,
+                    data_dimensions[1],
+                    data_dimensions[2]-1,
                     2)
                 # NOTE: Flattening happening in
                 # convert_to_dmap_datastructures
@@ -903,6 +906,17 @@ class BorealisConvert(BorealisRead):
                 lp_sw = 0
             else:
                 lp_sw = record_dict['lp_status_word']
+
+            # TX Antenna Mag only introduced in Borealis v0.7 onwards, so txpow defaults to -1 if not present.
+            # If present, txpow is a bitfield mapping of whether each antenna was transmitting. Antenna 15 is the
+            # MSB, and Antenna 0 the LSB. Since txpow is a signed int in DMAP, -1 means all antennas transmitting.
+            if 'tx_antenna_phases' not in record_dict.keys():
+                txpow = -1      # This is the same as if all antennas were transmitting.
+            else:
+                txpow = np.uint16()
+                for i in range(len(record_dict['tx_antenna_phases'])):
+                    if np.abs(record_dict['tx_antenna_phases'][i]) > 0:
+                        txpow += 1 << i
 
             sdarn_record_dict = {
                 'radar.revision.major': np.int8(borealis_major_revision),
@@ -945,7 +959,7 @@ class BorealisConvert(BorealisRead):
                                     utcfromtimestamp(
                                         record_dict['sqn_timestamps'][0]).
                                     microsecond),
-                'txpow': np.int16(-1),
+                'txpow': np.int16(txpow),
                 # see Borealis documentation
                 'nave': np.int16(record_dict['num_sequences']),
                 'atten': np.int16(0),
@@ -974,7 +988,7 @@ class BorealisConvert(BorealisRead):
                 'mppul': np.int16(len(record_dict['pulses'])),
                 # an alternate lag-zero will be given.
                 'mplgs': np.int16(record_dict['lags'].shape[0] - 1),
-                'nrang': np.int16(record_dict['correlation_dimensions'][1]),
+                'nrang': np.int16(data_dimensions[1]),
                 'frang': np.int16(round(record_dict['first_range'])),
                 'rsep': np.int16(round(record_dict['range_sep'])),
                 # False if list is empty.
@@ -997,7 +1011,7 @@ class BorealisConvert(BorealisRead):
                 'pwr0': lag_zero_power.astype(np.float32),
                 # list from 0 to num_ranges
                 'slist': np.array(list(
-                            range(0, record_dict['correlation_dimensions'][1]))
+                            range(0, data_dimensions[1]))
                             ).astype(np.int16),
                 'acfd': correlation_dict['main_acfs'],
                 'xcfd': correlation_dict['xcfs']
