@@ -11,6 +11,10 @@ BorealisBfiq
 BorealisAntennasIq
 BorealisRawrf
 as well as previous versions of these classes, currently including
+BorealisRawacfv0_6
+BorealisBfiqv0_6
+BorealisAntennasIqv0_6
+BorealisRawrfv0_6
 BorealisRawacfv0_5
 BorealisBfiqv0_5
 BorealisAntennasIqv0_5
@@ -55,6 +59,537 @@ from collections import OrderedDict
 from .base_format import BaseFormat
 
 
+class BorealisFieldsv0_4():
+    """
+    Class containing the mapping of Borealis data fields and types for each 
+    Borealis file type for Borealis v0.4 and earlier.
+
+    See Also
+    --------
+    BorealisFields (most up to date format)
+    """
+
+    @classmethod
+    def files_with_fields(cls):
+        """
+        Get the mapping of Borealis data fields to each file type.
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and a list of file types
+        as the values.
+        """
+        return {
+            "antenna_arrays_order": ['antennas_iq', 'bfiq'],
+            "beam_azms": ['antennas_iq', 'bfiq', 'rawacf'],
+            "beam_nums": ['antennas_iq', 'bfiq', 'rawacf'],
+            "blanked_samples": ['bfiq', 'rawacf'],
+            "borealis_git_hash": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "correlation_descriptors": ['rawacf'],
+            "correlation_dimensions": ['rawacf'],
+            "data": ['antennas_iq', 'bfiq', 'rawrf'],
+            "data_descriptors": ['antennas_iq', 'bfiq', 'rawrf'],
+            "data_dimensions": ['antennas_iq', 'bfiq', 'rawrf'],
+            "data_normalization_factor": ['antennas_iq', 'bfiq', 'rawacf'],
+            "experiment_id": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "experiment_name": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "experiment_comment": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "first_range": ['bfiq', 'rawacf'],
+            "first_range_rtt": ['bfiq', 'rawacf'],
+            "freq": ['antennas_iq', 'bfiq', 'rawacf'],
+            "int_time": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "intf_acfs": ['rawacf'],
+            "intf_antenna_count": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "lags": ['bfiq', 'rawacf'],
+            "main_acfs": ['rawacf'],
+            "main_antenna_count": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "noise_at_freq": ['antennas_iq', 'bfiq', 'rawacf'],
+            "num_beams": ['antennas_iq', 'bfiq', 'rawacf'],
+            "num_ranges": ['bfiq'],
+            "num_samps": ['antennas_iq', 'bfiq', 'rawrf'],
+            "num_sequences": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "num_slices": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "pulse_phase_offset": ['antennas_iq', 'bfiq'],
+            "pulses": ['antennas_iq', 'bfiq', 'rawacf'],
+            "range_sep": ['bfiq', 'rawacf'],
+            "rx_center_freq": ['rawrf'],
+            "rx_sample_rate": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "samples_data_type": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "scan_start_marker": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "slice_comment": ['antennas_iq', 'bfiq', 'rawacf'],
+            "sqn_timestamps": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "station": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "tau_spacing": ['antennas_iq', 'bfiq', 'rawacf'],
+            "tx_pulse_len": ['antennas_iq', 'bfiq', 'rawacf'],
+            "xcfs": ['rawacf'],
+        }
+
+    @classmethod
+    def all_single_element_types(cls):
+        """
+        Get the mapping of Borealis data fields to its single element type.
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and the data field variable
+        types as values.
+        """
+        return {
+            # Identifies the version of Borealis that made this data. Necessary
+            # for all versions.
+            "borealis_git_hash": str,
+            # data normalization factor determined by the filter scaling in the
+            # decimation scheme.
+            "data_normalization_factor": np.float64,
+            # Comment about the whole experiment
+            "experiment_comment": str,
+            # Number used to identify experiment.
+            "experiment_id": np.int64,
+            # Name of the experiment file.
+            "experiment_name": str,
+            # Distance to first range in km.
+            "first_range": np.float32,
+            # Round trip time of flight to first range in microseconds.
+            "first_range_rtt": np.float32,
+            # The frequency used for this experiment slice in kHz.
+            "freq": np.uint32,
+            # Integration time in seconds.
+            "int_time": np.float32,
+            # Number of interferometer array antennas.
+            "intf_antenna_count": np.uint32,
+            # Number of main array antennas.
+            "main_antenna_count": np.uint32,
+            # number of beams calculated for the integration time.
+            "num_beams": np.uint32,
+            # Number of ranges to calculate correlations for.
+            "num_ranges": np.uint32,
+            # Number of samples in the sampling period.
+            "num_samps": np.uint32,
+            # Number of sampling periods in the integration time.
+            "num_sequences": np.int64,
+            # Number of slices in the experiment at this integration time.
+            "num_slices": np.int64,
+            # range gate separation (equivalent distance between samples), km.
+            "range_sep": np.float32,
+            # The center frequency of this data in kHz
+            "rx_center_freq": np.float64,
+            # Sampling rate of the samples being written to file in Hz.
+            "rx_sample_rate": np.float64,
+            # str denoting C data type of the samples included in the data
+            # array, such as 'complex float'.
+            "samples_data_type": str,
+            # Designates if the record is the first in a scan.
+            "scan_start_marker": np.uint8,
+            # Additional text comment that describes the slice.
+            "slice_comment": str,
+            # Three letter radar identifier.
+            "station": str,
+            # The minimum spacing between pulses, spacing between pulses is
+            # always a multiple of this in microseconds.
+            "tau_spacing": np.uint32,
+            # Length of the pulse in microseconds.
+            "tx_pulse_len": np.uint32,
+        }
+
+    @classmethod
+    def all_array_types(cls):
+        """
+        Get the mapping of Borealis array data fields to its type
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and the data field variable
+        types as values.
+        """
+        return {
+            # States what order the data is in. Describes the data layout.
+            "antenna_arrays_order": np.str_,
+            # A list of the beams azimuths for each beam in degrees.
+            "beam_azms": np.float64,
+            # A list of beam numbers used in this slice.
+            "beam_nums": np.uint32,
+            # Samples that have been blanked during TR switching.
+            "blanked_samples": np.uint32,
+            # Denotes what each acf/xcf dimension represents. = "num_beams",
+            # "num_ranges", "num_lags" in site rawacf files.
+            "correlation_descriptors": np.str_,
+            # The dimensions in which to reshape the acf/xcf data.
+            "correlation_dimensions": np.uint32,
+            # A contiguous set of samples (complex float) at given sample rate
+            "data": np.complex64,
+            # Denotes what each data dimension represents. =
+            # "num_antenna_arrays", "num_sequences", "num_beams", "num_samps"
+            # for site bfiq.
+            "data_descriptors": np.str_,
+            # The dimensions in which to reshape the data.
+            "data_dimensions": np.uint32,
+            # Interferometer array autocorrelations
+            "intf_acfs": np.complex64,
+            # The lags created from combined pulses.
+            "lags": np.uint32,
+            # Main array autocorrelations
+            "main_acfs": np.complex64,
+            # Noise at the receive frequency, should be an array
+            # (one value per sequence) (TODO units??) (TODO document
+            # FFT resolution bandwidth for this value, should be =
+            # output_sample rate?)
+            "noise_at_freq": np.float64,
+            # For pulse encoding phase, in degrees offset.
+            # Contains one phase offset per pulse in pulses.
+            "pulse_phase_offset": np.float32,
+            # The pulse sequence in multiples of the tau_spacing.
+            "pulses": np.uint32,
+            # A list of GPS timestamps of the beginning of transmission for
+            # each sampling period in the integration time. Seconds since
+            # epoch.
+            "sqn_timestamps": np.float64,
+            # Crosscorrelations between main and interferometer arrays
+            "xcfs": np.complex64,
+        }
+
+    @classmethod
+    def single_element_types(cls, file_type: str) -> dict:
+        """
+        Gets the single element types of a given file type.
+
+        Parameters
+        ----------
+        file_type: str
+            File type to get single element fields for. One of 'antennas_iq', 
+            'bfiq', 'rawacf', or 'rawrf'
+
+        Returns
+        -------
+        dict
+            Dictionary of field: type for all fields contained in file_type.
+        """
+        relevant_fields = [k for k, v in cls.files_with_fields().items() if file_type in v]
+        single_elements = cls.all_single_element_types()
+        return {k: v for k, v in single_elements.items() if k in relevant_fields}
+
+    @classmethod
+    def array_types(cls, file_type: str) -> dict:
+        """
+        Gets the array types of a given file type.
+
+        Parameters
+        ----------
+        file_type: str
+            File type to get array fields for. One of 'antennas_iq', 'bfiq', 
+            'rawacf', or 'rawrf'
+
+        Returns
+        -------
+        dict
+            Dictionary of field: type for all fields contained in file_type.
+        """
+        relevant_fields = [k for k, v in cls.files_with_fields().items() if file_type in v]
+        array_elements = cls.all_array_types()
+        return {k: v for k, v in array_elements.items() if k in relevant_fields}
+
+    @classmethod
+    def all_shared_fields(cls):
+        """
+        List of all fields that are shared between records in a site-structured file.
+
+        Notes
+        -----
+        The dimension info for shared_fields is not necessary because the
+        dimensions will be the same for site and restructured files.
+        """
+        return ['antenna_arrays_order',
+                'blanked_samples',
+                'borealis_git_hash',
+                'data_normalization_factor',
+                'experiment_comment',
+                'experiment_id',
+                'experiment_name',
+                'first_range',
+                'first_range_rtt',
+                'freq',
+                'intf_antenna_count',
+                'lags',
+                'main_antenna_count',
+                'num_ranges',
+                'num_samps',
+                'pulse_phase_offset',
+                'pulses',
+                'range_sep',
+                'rx_sample_rate',
+                'samples_data_type',
+                'slice_comment',
+                'station',
+                'tau_spacing',
+                'tx_pulse_len']
+
+    @classmethod
+    def shared_fields(cls, file_type: str) -> list:
+        """
+        Gets the shared fields of a given file type.
+
+        Parameters
+        ----------
+        file_type: str
+            File type to get shared fields for. One of 'antennas_iq', 'bfiq',
+            'rawacf', or 'rawrf'
+
+        Returns
+        -------
+        list[str]
+            List of shared fields for the given data type.
+        """
+        relevant_fields = [k for k, v in cls.files_with_fields().items() if file_type in v]
+        shared = cls.all_shared_fields()
+        return [k for k in shared if k in relevant_fields]
+
+
+class BorealisFieldsv0_5(BorealisFieldsv0_4):
+    """
+    Class containing the mapping of Borealis data fields and types for each 
+    Borealis file type for Borealis v0.5.
+
+    See Also
+    --------
+    BorealisFields (most up to date format)
+    """
+
+    @classmethod
+    def files_with_fields(cls):
+        """
+        Get the mapping of Borealis data fields to each file type.
+        Mapping is updated to reflect changes from previous version of Borealis.
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and a list of file types
+        as the values.
+        """
+        field_file_mapping = super().files_with_fields()
+        field_file_mapping.update({
+            "averaging_method": ['rawacf'],
+            "num_blanked_samples": ['antennas_iq', 'bfiq', 'rawacf'],
+            "scheduling_mode": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "slice_id": ['antennas_iq', 'bfiq', 'rawacf'],
+            "slice_interfacing": ['antennas_iq', 'bfiq', 'rawacf'],
+        })
+        field_file_mapping['blanked_samples'].extend(['antennas_iq', 'rawrf'])
+
+        return field_file_mapping
+
+    @classmethod
+    def all_single_element_types(cls):
+        """
+        Get the mapping of Borealis data fields to its single element type.
+        Mapping is updated to reflect changes from previous version of Borealis.
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and the data field variable
+        types as values.
+        """
+        single_element_types = super().all_single_element_types()
+        single_element_types.update({
+            # A string describing the averaging method, ex. mean, median
+            "averaging_method": str,
+            # number of blanked samples in the sequence.
+            "num_blanked_samples": np.uint32,
+            # A string describing the type of scheduling time at the time of
+            # this dataset.
+            "scheduling_mode": str,
+            # the slice id of the file and dataset.
+            "slice_id": np.uint32,
+            # the interfacing of this slice to other slices.
+            "slice_interfacing": str,
+        })
+        return single_element_types
+
+    @classmethod
+    def all_shared_fields(cls):
+        """
+        List of all fields that are shared between records in a site-structured file.
+
+        Notes
+        -----
+        In Borealis v0.5, slice_id, scheduling_mode, and
+        averaging_method were added and these will be shared fields. These
+        fields will not change from record to record. Blanked samples may
+        change from record to record if a new slice is added and interfaced
+        within the sequence. Therefore, this bug was fixed by changing
+        blanked_samples to an unshared field in Borealis v0.5.
+        """
+        shared = super().all_shared_fields() + \
+            ['averaging_method', 'scheduling_mode', 'slice_id']
+        shared.remove('blanked_samples')
+        return shared
+
+
+class BorealisFieldsv0_6(BorealisFieldsv0_5):
+    """
+    Class containing the mapping of Borealis data fields and types for each 
+    Borealis file type for Borealis v0.6.
+
+    See Also
+    --------
+    BorealisFields (most up to date format)
+    """
+
+    @classmethod
+    def files_with_fields(cls):
+        """
+        Get the mapping of Borealis data fields to each file type.
+        Mapping is updated to reflect changes from previous version of Borealis.
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and a list of file types
+        as the values.
+        """
+        field_file_mapping = super().files_with_fields()
+        field_file_mapping.update({
+            "agc_status_word": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "gps_locked": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "gps_to_system_time_diff": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+            "lp_status_word": ['antennas_iq', 'bfiq', 'rawacf', 'rawrf'],
+        })
+
+        return field_file_mapping
+
+    @classmethod
+    def all_single_element_types(cls):
+        """
+        Get the mapping of Borealis data fields to its single element type.
+        Mapping is updated to reflect changes from previous version of Borealis.
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and the data field variable
+        types as values.
+        """
+        single_element_types = super().all_single_element_types()
+        single_element_types.update({
+            # the agc fault status of each transmitter, transmitter/USRP
+            # mapped to bit position
+            # A '1' indicates an agc fault at least once during the integration
+            # period.
+            "agc_status_word": np.uint32,
+            # Updated to 16 bit number to avoid mismatch when converting
+            # to DMAP format.
+            "experiment_id": np.int16,
+            # Boolean indicating if the GPS was locked during the entire
+            # integration period
+            "gps_locked": np.uint8,
+            # The max time diffe between GPS and system time during the
+            # integration period. In seconds. Negative if GPS time ahead.
+            "gps_to_system_time_diff": np.float64,
+            # the low power status of each transmitter, transmitter/USRP
+            # mapped to bit position
+            # A '1' indicates a low power condition at least once during the
+            # integration period.
+            "lp_status_word": np.uint32,
+        })
+        return single_element_types
+
+
+class BorealisFields(BorealisFieldsv0_6):
+    """
+    Class containing the mapping of Borealis data fields and types for each 
+    Borealis file type for the current version of Borealis
+    """
+
+    @classmethod
+    def files_with_fields(cls):
+        """
+        Get the mapping of Borealis data fields to each file type. 
+        Mapping is updated to reflect changes from previous version of Borealis.
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and a list of file types
+        as the values.
+        """
+        field_file_mapping = super().files_with_fields()
+        field_file_mapping.pop('correlation_descriptors')
+        field_file_mapping.pop('correlation_dimensions')
+        field_file_mapping['data_descriptors'].append('rawacf')
+        field_file_mapping['data_dimensions'].append('rawacf')
+        field_file_mapping['first_range'].append('antennas_iq')
+        field_file_mapping['first_range_rtt'].append('antennas_iq')
+        field_file_mapping['lags'].append('antennas_iq')
+        field_file_mapping['num_ranges'].append('antennas_iq')
+        field_file_mapping['range_sep'].append('antennas_iq')
+        field_file_mapping['tx_antenna_phases'] = ['antennas_iq', 'bfiq', 'rawacf', 'rawrf']
+
+        return field_file_mapping
+
+    @classmethod
+    def all_single_element_types(cls):
+        """
+        Get the mapping of Borealis data fields to its single element type.
+        Mapping is updated to reflect changes from previous version of Borealis.
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and the data field variable
+        types as values.
+
+        Notes
+        -----
+        In v0.7, gps_locked and scan_start_marker were changed to np.bool_ fields from np.uint8.
+        """
+        single_element_types = super().all_single_element_types()
+        single_element_types.update({
+            "gps_locked": np.bool_,
+            "scan_start_marker": np.bool_,
+        })
+        return single_element_types
+
+    @classmethod
+    def all_array_types(cls):
+        """
+        Get the mapping of Borealis array data fields to its type
+
+        Returns
+        -------
+        A dictionary containing data fields as keys and the data field variable
+        types as values.
+
+        Notes
+        -----
+        In v0.7, antenna_arrays_order and data_descriptors were changed to np.array(np.bytes_).
+        """
+        all_arrays = super().all_array_types()
+        all_arrays.update({
+            "antenna_arrays_order": np.bytes_,
+            "data_descriptors": np.bytes_,
+            "tx_antenna_phases": np.complex64,
+        })
+        all_arrays.pop('correlation_dimensions')
+        all_arrays.pop('correlation_descriptors')
+        return all_arrays
+
+    @staticmethod
+    def find_num_antennas_site(records: OrderedDict) -> int:
+        """
+        Find the number of antennas given the records dictionary, for restructuring to arrays.
+
+        Parameters
+        ----------
+        records
+            The records dictionary from a site-style file.
+
+        Returns
+        -------
+        num_antennas
+            The number of antennas that have been recorded and stored in the file.
+
+        Notes
+        -----
+        Num_antennas is unique to a slice so cannot change inside file.
+        """
+        first_key = list(records.keys())[0]
+        return records[first_key]['tx_antenna_phases'].shape[0]
+
+
 class BorealisRawacfv0_4(BaseFormat):
     """
     Class containing Borealis Rawacf data fields and their types.
@@ -81,6 +616,7 @@ class BorealisRawacfv0_4(BaseFormat):
     find_num_lags(OrderedDict): int
         Returns the num lags in the data for use in finding dimensions
     """
+    fields = BorealisFieldsv0_4
 
     @staticmethod
     def find_num_ranges(records: OrderedDict) -> int:
@@ -212,18 +748,14 @@ class BorealisRawacfv0_4(BaseFormat):
             List of parameter names that are not shared between all the records
             in the site restructured file, i.e. may have different dimensions
             between records.
+            
         Returns
         -------
         fields_max_dims: dict
             dictionary containing field names (str) as keys with maximum
             dimensions required to restructure to array file as values (tuples)
-        Raises
-        ------
-
         """
-        fields_max_dims, max_num_sequences, max_num_beams = super(BorealisRawacfv0_4,
-                                                                  cls).site_get_max_dims(filename,
-                                                                                         unshared_parameters)
+        fields_max_dims, max_num_sequences, max_num_beams = super().site_get_max_dims(filename, unshared_parameters)
 
         # Now change the main_acfs, int_acfs and xcfs dicts to maximum required dims
 
@@ -259,56 +791,7 @@ class BorealisRawacfv0_4(BaseFormat):
             All the single-element fields in records of the
             format, as a dictionary fieldname : type.
         """
-        return {
-            # Identifies the version of Borealis that made this data. Necessary
-            # for all versions.
-            "borealis_git_hash": np.unicode_,
-            # Number used to identify experiment.
-            "experiment_id": np.int64,
-            # Name of the experiment file.
-            "experiment_name": np.unicode_,
-            # Comment about the whole experiment
-            "experiment_comment": np.unicode_,
-            # Additional text comment that describes the slice.
-            "slice_comment": np.unicode_,
-            # Number of slices in the experiment at this integration time.
-            "num_slices": np.int64,
-            # Three letter radar identifier.
-            "station": np.unicode_,
-            # Number of sampling periods in the integration time.
-            "num_sequences": np.int64,
-            # range gate separation (equivalent distance between samples), km.
-            "range_sep": np.float32,
-            # Round trip time of flight to first range in microseconds.
-            "first_range_rtt": np.float32,
-            # Distance to first range in km.
-            "first_range": np.float32,
-            # Sampling rate of the samples being written to file in Hz.
-            "rx_sample_rate": np.float64,
-            # Designates if the record is the first in a scan.
-            "scan_start_marker": np.bool_,
-            # Integration time in seconds.
-            "int_time": np.float32,
-            # Length of the pulse in microseconds.
-            "tx_pulse_len": np.uint32,
-            # The minimum spacing between pulses, spacing between pulses is
-            # always a multiple of this in microseconds.
-            "tau_spacing": np.uint32,
-            # Number of main array antennas.
-            "main_antenna_count": np.uint32,
-            # Number of interferometer array antennas.
-            "intf_antenna_count": np.uint32,
-            # The frequency used for this experiment slice in kHz.
-            "freq": np.uint32,
-            # str denoting C data type of the samples included in the data
-            # array, such as 'complex float'.
-            "samples_data_type": np.unicode_,
-            # data normalization factor determined by the filter scaling in the
-            # decimation scheme.
-            "data_normalization_factor": np.float64,
-            # number of beams calculated for the integration time.
-            "num_beams": np.uint32
-            }
+        return cls.fields.single_element_types('rawacf')
 
     @classmethod
     def array_dtypes(cls):
@@ -321,38 +804,7 @@ class BorealisRawacfv0_4(BaseFormat):
             All the array fields in records of the
             format, as a dictionary fieldname : array dtype.
         """
-        return {
-            # The pulse sequence in multiples of the tau_spacing.
-            "pulses": np.uint32,
-            # The lags created from combined pulses.
-            "lags": np.uint32,
-            # Samples that have been blanked during TR switching.
-            "blanked_samples": np.uint32,
-            # A list of GPS timestamps of the beginning of transmission for
-            # each sampling period in the integration time. Seconds since
-            # epoch.
-            "sqn_timestamps": np.float64,
-            # A list of beam numbers used in this slice.
-            "beam_nums": np.uint32,
-            # A list of the beams azimuths for each beam in degrees.
-            "beam_azms": np.float64,
-            # Noise at the receive frequency, should be an array
-            # (one value per sequence) (TODO units??) (TODO document
-            # FFT resolution bandwidth for this value, should be =
-            # output_sample rate?)
-            "noise_at_freq": np.float64,
-            # Denotes what each acf/xcf dimension represents. = "num_beams",
-            # "num_ranges", "num_lags" in site rawacf files.
-            "correlation_descriptors": np.unicode_,
-            # The dimensions in which to reshape the acf/xcf data.
-            "correlation_dimensions": np.uint32,
-            # Main array autocorrelations
-            "main_acfs": np.complex64,
-            # Interferometer array autocorrelations
-            "intf_acfs": np.complex64,
-            # Crosscorrelations between main and interferometer arrays
-            "xcfs": np.complex64
-            }
+        return cls.fields.array_types('rawacf')
 
     @classmethod
     def shared_fields(cls):
@@ -364,14 +816,7 @@ class BorealisRawacfv0_4(BaseFormat):
         The dimension info for shared_fields is not necessary because the
         dimensions will be the same for site and restructured files.
         """
-        return ['blanked_samples', 'borealis_git_hash',
-                'data_normalization_factor', 'experiment_comment',
-                'experiment_id', 'experiment_name', 'first_range',
-                'first_range_rtt', 'freq', 'intf_antenna_count', 'lags',
-                'main_antenna_count', 'pulses', 'range_sep',
-                'rx_sample_rate', 'samples_data_type',
-                'slice_comment', 'station', 'tau_spacing',
-                'tx_pulse_len']
+        return cls.fields.shared_fields('rawacf')
 
     @classmethod
     def unshared_fields_dims_array(cls):
@@ -496,6 +941,7 @@ class BorealisBfiqv0_4(BaseFormat):
     find_num_samps(OrderedDict): int
         Returns the number of samples in the data for use in finding dimensions
     """
+    fields = BorealisFieldsv0_4
 
     @staticmethod
     def find_num_antenna_arrays(records: OrderedDict) -> int:
@@ -626,18 +1072,14 @@ class BorealisBfiqv0_4(BaseFormat):
             List of parameter names that are not shared between all the records
             in the site restructured file, i.e. may have different dimensions
             between records.
+
         Returns
         -------
         fields_max_dims: dict
             dictionary containing field names (str) as keys with maximum
             dimensions required to restructure to array file as values (tuples)
-        Raises
-        ------
-
         """
-        fields_max_dims, max_num_sequences, max_num_beams = super(BorealisBfiqv0_4,
-                                                                  cls).site_get_max_dims(filename,
-                                                                                         unshared_parameters)
+        fields_max_dims, max_num_sequences, max_num_beams = super().site_get_max_dims(filename, unshared_parameters)
 
         # Get the num_ant_arrays and num_samps fields directly from one record of the file
         with h5py.File(filename, 'r') as site_file:
@@ -669,60 +1111,7 @@ class BorealisBfiqv0_4(BaseFormat):
             All the single-element fields in records of the
             format, as a dictionary fieldname : type.
         """
-        return {
-            # Identifies the version of Borealis that made this data. Necessary
-            # for all versions.
-            "borealis_git_hash": np.unicode_,
-            # Number used to identify experiment.
-            "experiment_id": np.int64,
-            # Name of the experiment file.
-            "experiment_name": np.unicode_,
-            # Comment about the whole experiment
-            "experiment_comment": np.unicode_,
-            # Additional text comment that describes the slice.
-            "slice_comment": np.unicode_,
-            # Number of slices in the experiment at this integration time.
-            "num_slices": np.int64,
-            # Three letter radar identifier.
-            "station": np.unicode_,
-            # Number of sampling periods in the integration time.
-            "num_sequences": np.int64,
-            # Sampling rate of the samples being written to file in Hz.
-            "rx_sample_rate": np.float64,
-            # Designates if the record is the first in a scan.
-            "scan_start_marker": np.bool_,
-            # Integration time in seconds.
-            "int_time": np.float32,
-            # Length of the pulse in microseconds.
-            "tx_pulse_len": np.uint32,
-            # The minimum spacing between pulses, spacing between pulses is
-            # always a multiple of this. In microseconds.
-            "tau_spacing": np.uint32,
-            # Number of main array antennas.
-            "main_antenna_count": np.uint32,
-            # Number of interferometer array antennas.
-            "intf_antenna_count": np.uint32,
-            # The frequency used for this experiment slice in kHz.
-            "freq": np.uint32,
-            # str denoting C data type of the samples included in the data
-            # array, such as 'complex float'.
-            "samples_data_type": np.unicode_,
-            # Number of samples in the sampling period.
-            "num_samps": np.uint32,
-            # range gate separation (equivalent distance between samples), km
-            "range_sep": np.float32,
-            # Round trip time of flight to first range in microseconds.
-            "first_range_rtt": np.float32,
-            # Distance to first range in km.
-            "first_range": np.float32,
-            # Number of ranges to calculate correlations for.
-            "num_ranges": np.uint32,
-            # data normalization factor determined by the filter scaling in the
-            # decimation scheme.
-            "data_normalization_factor": np.float64,
-            # number of beams calculated for the integration time.
-            "num_beams": np.uint32
-            }
+        return cls.fields.single_element_types('bfiq')
 
     @classmethod
     def array_dtypes(cls):
@@ -735,56 +1124,19 @@ class BorealisBfiqv0_4(BaseFormat):
             All the array fields in records of the
             format, as a dictionary fieldname : array dtype.
         """
-        return {
-            # The pulse sequence in multiples of the tau_spacing.
-            "pulses": np.uint32,
-            # The lags created from combined pulses.
-            "lags": np.uint32,
-            # Samples that have been blanked during TR switching.
-            "blanked_samples": np.uint32,
-            # For pulse encoding phase, in degrees offset.
-            # Contains one phase offset per pulse in pulses.
-            "pulse_phase_offset": np.float32,
-            # A list of GPS timestamps of the beginning of transmission for
-            # each sampling period in the integration time. Seconds since
-            # epoch.
-            "sqn_timestamps": np.float64,
-            # A list of beam numbers used in this slice.
-            "beam_nums": np.uint32,
-            # A list of the beams azimuths for each beam in degrees.
-            "beam_azms": np.float64,
-            # Noise at the receive frequency, should be an array (one value per
-            # sequence) (TODO units??) (TODO document FFT resolution
-            # bandwidth for this value, should be = output_sample rate?)
-            "noise_at_freq": np.float64,
-            # States what order the data is in. Describes the data layout.
-            "antenna_arrays_order": np.unicode_,
-            # Denotes what each data dimension represents. =
-            # "num_antenna_arrays", "num_sequences", "num_beams", "num_samps"
-            # for site bfiq.
-            "data_descriptors": np.unicode_,
-            # The dimensions in which to reshape the data.
-            "data_dimensions": np.uint32,
-            # A contiguous set of samples (complex float) at given sample rate
-            "data": np.complex64
-            }
+        return cls.fields.array_types('bfiq')
 
     @classmethod
     def shared_fields(cls):
         """
         See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        The dimension info for shared_fields is not necessary because the
+        dimensions will be the same for site and restructured files.
         """
-        return ['antenna_arrays_order', 'blanked_samples',
-                'borealis_git_hash',
-                'data_normalization_factor',
-                'experiment_comment', 'experiment_id', 'experiment_name',
-                'first_range', 'first_range_rtt', 'freq',
-                'intf_antenna_count', 'lags', 'main_antenna_count',
-                'num_ranges', 'num_samps',
-                'pulse_phase_offset', 'pulses', 'range_sep',
-                'rx_sample_rate', 'samples_data_type',
-                'slice_comment', 'station', 'tau_spacing',
-                'tx_pulse_len']
+        return cls.fields.shared_fields('bfiq')
 
     @classmethod
     def unshared_fields_dims_array(cls):
@@ -896,6 +1248,7 @@ class BorealisAntennasIqv0_4(BaseFormat):
     find_num_samps(OrderedDict): int
         Returns the number of samples in the data for use in finding dimensions
     """
+    fields = BorealisFieldsv0_4
 
     @staticmethod
     def find_num_antennas(records: OrderedDict) -> int:
@@ -981,8 +1334,6 @@ class BorealisAntennasIqv0_4(BaseFormat):
 
         return new_records
 
-
-
     @staticmethod
     def flatten_site_arrays(records: OrderedDict) -> OrderedDict:
         """
@@ -1037,9 +1388,7 @@ class BorealisAntennasIqv0_4(BaseFormat):
         ------
 
         """
-        fields_max_dims, max_num_sequences, max_num_beams = super(BorealisAntennasIqv0_4,
-                                                                  cls).site_get_max_dims(filename,
-                                                                                         unshared_parameters)
+        fields_max_dims, max_num_sequences, max_num_beams = super().site_get_max_dims(filename, unshared_parameters)
 
         # Get the num_antennas and num_samps fields directly from one record of the file
         with h5py.File(filename, 'r') as site_file:
@@ -1071,52 +1420,7 @@ class BorealisAntennasIqv0_4(BaseFormat):
             All the single-element fields in records of the
             format, as a dictionary fieldname : type.
         """
-        return {
-            # Identifies the version of Borealis that made this data. Necessary
-            # for all versions.
-            "borealis_git_hash": np.unicode_,
-            # Number used to identify experiment.
-            "experiment_id": np.int64,
-            # Name of the experiment file.
-            "experiment_name": np.unicode_,
-            # Comment about the whole experiment
-            "experiment_comment": np.unicode_,
-            # Additional text comment that describes the slice.
-            "slice_comment": np.unicode_,
-            # Number of slices in the experiment at this integration time.
-            "num_slices": np.int64,
-            # Three letter radar identifier.
-            "station": np.unicode_,
-            # Number of sampling periods in the integration time.
-            "num_sequences": np.int64,
-            # Sampling rate of the samples being written to file in Hz.
-            "rx_sample_rate": np.float64,
-            # Designates if the record is the first in a scan.
-            "scan_start_marker": np.bool_,
-            # Integration time in seconds.
-            "int_time": np.float32,
-            # Length of the pulse in microseconds.
-            "tx_pulse_len": np.uint32,
-            # The minimum spacing between pulses, spacing between pulses is
-            # always a multiple of this in microseconds.
-            "tau_spacing": np.uint32,
-            # Number of main array antennas.
-            "main_antenna_count": np.uint32,
-            # Number of interferometer array antennas.
-            "intf_antenna_count": np.uint32,
-            # The frequency used for this experiment slice in kHz.
-            "freq": np.uint32,
-            # str denoting C data type of the samples included in the data
-            # array, such as 'complex float'.
-            "samples_data_type": np.unicode_,
-            # Number of samples in the sampling period.
-            "num_samps": np.uint32,
-            # data normalization factor determined by the filter scaling in the
-            # decimation scheme.
-            "data_normalization_factor": np.float64,
-            # number of beams to be calculated for the integration time.
-            "num_beams": np.uint32
-            }
+        return cls.fields.single_element_types('antennas_iq')
 
     @classmethod
     def array_dtypes(cls):
@@ -1129,49 +1433,19 @@ class BorealisAntennasIqv0_4(BaseFormat):
             All the array fields in records of the
             format, as a dictionary fieldname : array dtype.
         """
-        return {
-            # The pulse sequence in multiples of the tau_spacing.
-            "pulses": np.uint32,
-            # For pulse encoding phase, in desgrees offset. Contains one phase
-            # offset per pulse in pulses.
-            "pulse_phase_offset": np.float32,
-            # A list of GPS timestamps of the beginning of transmission for
-            # each sampling period in the integration time. Seconds since
-            # epoch.
-            "sqn_timestamps": np.float64,
-            # A list of beam numbers used in this slice.
-            "beam_nums": np.uint32,
-            # A list of the beams azimuths for each beam in degrees.
-            "beam_azms": np.float64,
-            # Noise at the receive frequency, should be an array (one value
-            # per sequence) (TODO units??) (TODO document FFT resolution
-            # bandwidth for this value, should be = output_sample rate?)
-            "noise_at_freq": np.float64,
-            # States what order the data is in. Describes the data layout.
-            "antenna_arrays_order": np.unicode_,
-            # Denotes what each data dimension represents. = "num_antennas",
-            # "num_sequences", "num_samps" in site antennas_iq.
-            "data_descriptors": np.unicode_,
-            # The dimensions in which to reshape the data.
-            "data_dimensions": np.uint32,
-            # A contiguous set of samples (complex float) at given sample rate
-            "data": np.complex64
-            }
+        return cls.fields.array_types('antennas_iq')
 
     @classmethod
     def shared_fields(cls):
         """
         See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        The dimension info for shared_fields is not necessary because the
+        dimensions will be the same for site and restructured files.
         """
-        return ['antenna_arrays_order',
-                'borealis_git_hash',
-                'data_normalization_factor', 'experiment_comment',
-                'experiment_id', 'experiment_name', 'freq',
-                'intf_antenna_count', 'main_antenna_count', 'num_samps',
-                'pulse_phase_offset', 'pulses',
-                'rx_sample_rate', 'samples_data_type',
-                'slice_comment', 'station', 'tau_spacing',
-                'tx_pulse_len']
+        return cls.fields.shared_fields('antennas_iq')
 
     @classmethod
     def unshared_fields_dims_array(cls):
@@ -1272,6 +1546,7 @@ class BorealisRawrfv0_4(BaseFormat):
     and how they are used to verify format files and restructure Borealis
     files to array and site structure.
     """
+    fields = BorealisFieldsv0_4
 
     @staticmethod
     def reshape_site_arrays(records: OrderedDict) -> OrderedDict:
@@ -1351,18 +1626,14 @@ class BorealisRawrfv0_4(BaseFormat):
             List of parameter names that are not shared between all the records
             in the site restructured file, i.e. may have different dimensions
             between records.
+
         Returns
         -------
         fields_max_dims: dict
             dictionary containing field names (str) as keys with maximum
             dimensions required to restructure to array file as values (tuples)
-        Raises
-        ------
-
         """
-        fields_max_dims, max_num_sequences, max_num_beams = super(BorealisRawrfv0_4,
-                                                                  cls).site_get_max_dims(filename,
-                                                                                         unshared_parameters)
+        fields_max_dims, max_num_sequences, max_num_beams = super().site_get_max_dims(filename, unshared_parameters)
 
         # Get the num_antennas and num_samps fields directly from one record of the file
         with h5py.File(filename, 'r') as site_file:
@@ -1399,40 +1670,7 @@ class BorealisRawrfv0_4(BaseFormat):
             All the single-element fields in records of the
             format, as a dictionary fieldname : type.
         """
-        return {
-            # Identifies the version of Borealis that made this data. Necessary
-            # for all versions.
-            "borealis_git_hash": np.unicode_,
-            # Number used to identify experiment.
-            "experiment_id": np.int64,
-            # Name of the experiment file.
-            "experiment_name": np.unicode_,
-            # Comment about the whole experiment
-            "experiment_comment": np.unicode_,
-            # Number of slices in the experiment at this integration time.
-            "num_slices": np.int64,
-            # Three letter radar identifier.
-            "station": np.unicode_,
-            # Number of sampling periods in the integration time.
-            "num_sequences": np.int64,
-            # Sampling rate of the samples being written to file in Hz.
-            "rx_sample_rate": np.float64,
-            # Designates if the record is the first in a scan.
-            "scan_start_marker": np.bool_,
-            # Integration time in seconds.
-            "int_time": np.float32,
-            # Number of main array antennas.
-            "main_antenna_count": np.uint32,
-            # Number of interferometer array antennas.
-            "intf_antenna_count": np.uint32,
-            # str denoting C data type of the samples included in the data
-            # array, such as 'complex float'.
-            "samples_data_type": np.unicode_,
-            # The center frequency of this data in kHz
-            "rx_center_freq": np.float64,
-            # Number of samples in the sampling period.
-            "num_samps": np.uint32
-            }
+        return cls.fields.single_element_types('rawrf')
 
     @classmethod
     def array_dtypes(cls):
@@ -1445,19 +1683,7 @@ class BorealisRawrfv0_4(BaseFormat):
             All the array fields in records of the
             format, as a dictionary fieldname : array dtype.
         """
-        return {
-            # A list of GPS timestamps of the beginning of transmission for
-            # each sampling period in the integration time. Seconds since
-            # epoch.
-            "sqn_timestamps": np.float64,
-            # Denotes what each data dimension represents. = "num_sequences",
-            # "num_antennas", "num_samps" for rawrf.
-            "data_descriptors": np.unicode_,
-            # The dimensions in which to reshape the data.
-            "data_dimensions": np.uint32,
-            # A contiguous set of samples (complex float) at given sample rate
-            "data": np.complex64
-            }
+        return cls.fields.array_types('rawrf')
 
 
 class BorealisRawacfv0_5(BorealisRawacfv0_4):
@@ -1487,61 +1713,7 @@ class BorealisRawacfv0_5(BorealisRawacfv0_4):
     num_blanked_samples, to specify how much data to read in the
     blanked_samples array in the array style file.
     """
-
-    @classmethod
-    def single_element_types(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Returns
-        -------
-        single_element_types
-            All the single-element fields in records of the
-            format, as a dictionary fieldname : type.
-
-        Notes
-        -----
-        In Borealis v0.5, slice_id, slice_interfacing, scheduling_mode, and
-        averaging_methods were added as single element fields.
-        num_blanked_samples is an array-specific field as the number may vary
-        from record to record, but it is also a single-element-per-record
-        field.
-        """
-        single_element_types = super(BorealisRawacfv0_5,
-                                     cls).single_element_types()
-        single_element_types.update({
-            # the slice id of the file and dataset.
-            "slice_id": np.uint32,
-            # the interfacing of this slice to other slices.
-            "slice_interfacing": np.unicode_,
-            # A string describing the type of scheduling time at the time of
-            # this dataset.
-            "scheduling_mode": np.unicode_,
-            # A string describing the averaging method, ex. mean, median
-            "averaging_method": np.unicode_,
-            # number of blanked samples in the sequence.
-            "num_blanked_samples": np.uint32
-            })
-        return single_element_types
-
-    @classmethod
-    def shared_fields(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Notes
-        -----
-        In Borealis v0.5, slice_id, scheduling_mode, and
-        averaging_method were added and these will be shared fields. These
-        fields will not change from record to record. Blanked samples may
-        change from record to record if a new slice is added and interfaced
-        within the sequence. Therefore, this bug was fixed by changing
-        blanked_samples to an unshared field in Borealis v0.5.
-        """
-        shared = super(BorealisRawacfv0_5, cls).shared_fields() + \
-            ['slice_id', 'scheduling_mode', 'averaging_method']
-        shared.remove('blanked_samples')
-        return shared
+    fields = BorealisFieldsv0_5
 
     @classmethod
     def unshared_fields_dims_array(cls):
@@ -1556,8 +1728,7 @@ class BorealisRawacfv0_5(BorealisRawacfv0_4):
         new slices may be added and interfaced to this slice and therefore
         slice_interfacing may not be the same from record to record.
         """
-        unshared_fields_dims = super(BorealisRawacfv0_5,
-                                     cls).unshared_fields_dims_array()
+        unshared_fields_dims = super().unshared_fields_dims_array()
         unshared_fields_dims.update({
             'blanked_samples': [cls.
                                 find_max_field_len_func('blanked_samples')],
@@ -1570,8 +1741,7 @@ class BorealisRawacfv0_5(BorealisRawacfv0_4):
         """
         See BaseFormat class for description and use of this method.
         """
-        unshared_fields_dims = super(BorealisRawacfv0_5,
-                                     cls).unshared_fields_dims_site()
+        unshared_fields_dims = super().unshared_fields_dims_site()
         unshared_fields_dims.update({
             'blanked_samples': [lambda arrays, record_num:
                                 arrays['num_blanked_samples'][record_num]],
@@ -1585,8 +1755,7 @@ class BorealisRawacfv0_5(BorealisRawacfv0_4):
         """
         See BaseFormat class for description and use of this method.
         """
-        array_specific = super(BorealisRawacfv0_5,
-                               cls).array_specific_fields_generate()
+        array_specific = super().array_specific_fields_generate()
         array_specific.update({
             'num_blanked_samples': lambda records: np.array(
                 [len(record['blanked_samples']) for key, record in
@@ -1599,8 +1768,7 @@ class BorealisRawacfv0_5(BorealisRawacfv0_4):
         """
         See BaseFormat class for description and use of this method.
         """
-        array_specific = super(BorealisRawacfv0_5,
-                               cls).array_specific_fields_iterative_generator()
+        array_specific = super().array_specific_fields_iterative_generator()
         array_specific.update({
             'num_blanked_samples': lambda record: len(record['blanked_samples'])
             })
@@ -1634,62 +1802,7 @@ class BorealisBfiqv0_5(BorealisBfiqv0_4):
     num_blanked_samples, to specify how much data to read in the
     blanked_samples array in the array style file.
     """
-
-    @classmethod
-    def single_element_types(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Returns
-        -------
-        single_element_types
-            All the single-element fields in records of the
-            format, as a dictionary fieldname : type.
-
-        See Also
-        --------
-        https://borealis.readthedocs.io/en/latest/borealis_data.html
-
-        Notes
-        -----
-        In Borealis v0.5, slice_id, slice_interfacing, and scheduling_mode
-        were added as single element fields. num_blanked_samples is an
-        array-specific field as the number may vary from record to record,
-        but it is also a single-element-per-record field.
-        """
-        single_element_types = super(BorealisBfiqv0_5,
-                                     cls).single_element_types()
-        single_element_types.update({
-            # the slice id of the file and dataset.
-            "slice_id": np.uint32,
-            # the interfacing of this slice to other slices.
-            "slice_interfacing": np.unicode_,
-            # A string describing the type of scheduling time at the time of
-            # this dataset.
-            "scheduling_mode": np.unicode_,
-            # number of blanked samples in the sequence.
-            "num_blanked_samples": np.uint32
-            })
-        return single_element_types
-
-    @classmethod
-    def shared_fields(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Notes
-        -----
-        In Borealis v0.5, slice_id and scheduling_mode were added and these
-        will be shared fields. These fields will not change from record to
-        record. Blanked samples may  change from record to record if a new
-        slice is added and interfaced within the sequence. Therefore, this bug
-        was fixed by changing blanked_samples to an unshared field in Borealis
-        v0.5.
-        """
-        shared = super(BorealisBfiqv0_5, cls).shared_fields() + \
-            ['slice_id', 'scheduling_mode']
-        shared.remove('blanked_samples')
-        return shared
+    fields = BorealisFieldsv0_5
 
     @classmethod
     def unshared_fields_dims_array(cls):
@@ -1704,8 +1817,7 @@ class BorealisBfiqv0_5(BorealisBfiqv0_4):
         new slices may be added and interfaced to this slice and therefore
         slice_interfacing may not be the same from record to record.
         """
-        unshared_fields_dims = super(BorealisBfiqv0_5,
-                                     cls).unshared_fields_dims_array()
+        unshared_fields_dims = super().unshared_fields_dims_array()
         unshared_fields_dims.update({
             'blanked_samples': [cls.
                                 find_max_field_len_func('blanked_samples')],
@@ -1718,8 +1830,7 @@ class BorealisBfiqv0_5(BorealisBfiqv0_4):
         """
         See BaseFormat class for description and use of this method.
         """
-        unshared_fields_dims = super(BorealisBfiqv0_5,
-                                     cls).unshared_fields_dims_site()
+        unshared_fields_dims = super().unshared_fields_dims_site()
         unshared_fields_dims.update({
             'blanked_samples': [lambda arrays, record_num:
                                 arrays['num_blanked_samples'][record_num]],
@@ -1733,8 +1844,7 @@ class BorealisBfiqv0_5(BorealisBfiqv0_4):
         """
         See BaseFormat class for description and use of this method.
         """
-        array_specific = super(BorealisBfiqv0_5,
-                               cls).array_specific_fields_generate()
+        array_specific = super().array_specific_fields_generate()
         array_specific.update({
             'num_blanked_samples': lambda records: np.array(
                 [len(record['blanked_samples']) for key, record in
@@ -1747,8 +1857,7 @@ class BorealisBfiqv0_5(BorealisBfiqv0_4):
         """
         See BaseFormat class for description and use of this method.
         """
-        array_specific = super(BorealisBfiqv0_5,
-                               cls).array_specific_fields_iterative_generator()
+        array_specific = super().array_specific_fields_iterative_generator()
         array_specific.update({
             'num_blanked_samples': lambda record: len(record['blanked_samples'])
         })
@@ -1781,76 +1890,7 @@ class BorealisAntennasIqv0_5(BorealisAntennasIqv0_4):
     an array-specific field, num_blanked_samples, to specify how much data to
     read in the blanked_samples array in the array style file.
     """
-
-    @classmethod
-    def single_element_types(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Returns
-        -------
-        single_element_types
-            All the single-element fields in records of the
-            format, as a dictionary fieldname : type.
-
-        Notes
-        -----
-        In Borealis v0.5, slice_id, slice_interfacing, and scheduling_mode
-        were added as single element fields. num_blanked_samples is an
-        array-specific field as the number may vary from record to record,
-        but it is also a single-element-per-record field.
-        """
-        single_element_types = super(BorealisAntennasIqv0_5,
-                                     cls).single_element_types()
-        single_element_types.update({
-            # the slice id of the file and dataset.
-            "slice_id": np.uint32,
-            # the interfacing of this slice to other slices.
-            "slice_interfacing": np.unicode_,
-            # A string describing the type of scheduling time at the time of
-            # this dataset.
-            "scheduling_mode": np.unicode_,
-            # number of blanked samples in the sequence.
-            "num_blanked_samples": np.uint32
-            })
-        return single_element_types
-
-    @classmethod
-    def array_dtypes(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Returns
-        -------
-        array_dtypes
-            All the array fields in records of the
-            format, as a dictionary fieldname : array dtype.
-
-        Notes
-        -----
-        In Borealis v0.5, blanked_samples was added as an array
-        field.
-        """
-        array_dtypes = super(BorealisAntennasIqv0_5, cls).array_dtypes()
-        array_dtypes.update({
-            # Samples that occur during TR switching (transmission times)
-            "blanked_samples": np.uint32
-            })
-        return array_dtypes
-
-    @classmethod
-    def shared_fields(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Notes
-        -----
-        In Borealis v0.5, slice_id and scheduling_mode were added and these
-        will be shared fields.
-        """
-        shared = super(BorealisAntennasIqv0_5, cls).shared_fields() + \
-            ['slice_id', 'scheduling_mode']
-        return shared
+    fields = BorealisFieldsv0_5
 
     @classmethod
     def unshared_fields_dims_array(cls):
@@ -1865,8 +1905,7 @@ class BorealisAntennasIqv0_5(BorealisAntennasIqv0_4):
         new slices may be added and interfaced to this slice and therefore
         the field may not be the same from record to record.
         """
-        unshared_fields_dims = super(BorealisAntennasIqv0_5,
-                                     cls).unshared_fields_dims_array()
+        unshared_fields_dims = super().unshared_fields_dims_array()
         unshared_fields_dims.update({
             'blanked_samples': [cls.
                                 find_max_field_len_func('blanked_samples')],
@@ -1879,8 +1918,7 @@ class BorealisAntennasIqv0_5(BorealisAntennasIqv0_4):
         """
         See BaseFormat class for description and use of this method.
         """
-        unshared_fields_dims = super(BorealisAntennasIqv0_5,
-                                     cls).unshared_fields_dims_site()
+        unshared_fields_dims = super().unshared_fields_dims_site()
         unshared_fields_dims.update({
             'blanked_samples': [lambda arrays, record_num:
                                 arrays['num_blanked_samples'][record_num]],
@@ -1893,8 +1931,7 @@ class BorealisAntennasIqv0_5(BorealisAntennasIqv0_4):
         """
         See BaseFormat class for description and use of this method.
         """
-        array_specific = super(BorealisAntennasIqv0_5,
-                               cls).array_specific_fields_generate()
+        array_specific = super().array_specific_fields_generate()
         array_specific.update({
             'num_blanked_samples': lambda records: np.array(
                 [len(record['blanked_samples']) for key, record in
@@ -1907,8 +1944,7 @@ class BorealisAntennasIqv0_5(BorealisAntennasIqv0_4):
         """
         See BaseFormat class for description and use of this method.
         """
-        array_specific = super(BorealisAntennasIqv0_5,
-                               cls).array_specific_fields_iterative_generator()
+        array_specific = super().array_specific_fields_iterative_generator()
         array_specific.update({
             'num_blanked_samples': lambda record: len(record['blanked_samples'])
         })
@@ -1933,63 +1969,13 @@ class BorealisRawrfv0_5(BorealisRawrfv0_4):
     In v0.5, the following fields were added to BorealisRawrf:
     scheduling_mode and blanked_samples.
     """
-
-    @classmethod
-    def single_element_types(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Returns
-        -------
-        single_element_types
-            All the single-element fields in records of the
-            format, as a dictionary fieldname : type.
-
-        Notes
-        -----
-        In Borealis v0.5, scheduling_mode was added as a single
-        element field.
-        """
-        single_element_types = super(BorealisRawrfv0_5,
-                                     cls).single_element_types()
-        single_element_types.update({
-            # A string describing the type of scheduling time at the time of
-            # this dataset.
-            "scheduling_mode": np.unicode_
-            })
-        return single_element_types
-
-    @classmethod
-    def array_dtypes(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Returns
-        -------
-        array_dtypes
-            All the array fields in records of the
-            format, as a dictionary fieldname : array dtype.
-
-        Notes
-        -----
-        In Borealis v0.5, blanked_samples was added as an array
-        field.
-        """
-        array_dtypes = super(BorealisRawrfv0_5, cls).array_dtypes()
-        array_dtypes.update({
-            # Samples that occur during TR switching (transmission times)
-            "blanked_samples": np.uint32
-            })
-        return array_dtypes
+    fields = BorealisFieldsv0_5
 
 
-# The following are the currently used classes, with additions according
-# to Borealis updates.
-
-class BorealisRawacf(BorealisRawacfv0_5):
+class BorealisRawacfv0_6(BorealisRawacfv0_5):
     """
-    Class containing Borealis Rawacf data fields and their types for the
-    current version of Borealis (v0.6).
+    Class containing Borealis Rawacf data fields and their types for Borealis 
+    version 0.6.
 
     See Also
     --------
@@ -2012,47 +1998,7 @@ class BorealisRawacf(BorealisRawacfv0_5):
     lp_status_word. Array structured files contain the same fields,
     but with dims of [num_records].
     """
-
-    @classmethod
-    def single_element_types(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Returns
-        -------
-        single_element_types
-            All the single-element fields in records of the
-            format, as a dictionary fieldname : type.
-
-        Notes
-        -----
-        In Borealis v0.6, gps_locked, gps_to_system_time_diff, agc_status_word,
-        and lp_status_word were added as single element fields
-        """
-        single_element_types = super(BorealisRawacf,
-                                     cls).single_element_types()
-        single_element_types.update({
-            # the agc fault status of each transmitter, transmitter/USRP
-            # mapped to bit position
-            # A '1' indicates an agc fault at least once during the integration
-            # period.
-            "agc_status_word": np.uint32,
-            # the low power status of each transmitter, transmitter/USRP
-            # mapped to bit position
-            # A '1' indicates a low power condition at least once during the
-            # integration period.
-            "lp_status_word": np.uint32,
-            # Boolean indicating if the GPS was locked during the entire
-            # integration period
-            "gps_locked": np.bool_,
-            # The max time diffe between GPS and system time during the
-            # integration period. In seconds. Negative if GPS time ahead.
-            "gps_to_system_time_diff": np.float64,
-            # Updated to 16 bit number to avoid mismatch when converting
-            # to DMAP format.
-            "experiment_id": np.int16
-        })
-        return single_element_types
+    fields = BorealisFieldsv0_6
 
     @classmethod
     def unshared_fields_dims_array(cls):
@@ -2066,8 +2012,7 @@ class BorealisRawacf(BorealisRawacfv0_5):
         All are unshared fields because their values may not be the same from
         record to record.
         """
-        unshared_fields_dims = super(BorealisRawacf,
-                                     cls).unshared_fields_dims_array()
+        unshared_fields_dims = super().unshared_fields_dims_array()
         unshared_fields_dims.update({
             'agc_status_word': [],
             'lp_status_word': [],
@@ -2081,8 +2026,7 @@ class BorealisRawacf(BorealisRawacfv0_5):
         """
         See BaseFormat class for description and use of this method.
         """
-        unshared_fields_dims = super(BorealisRawacf,
-                                     cls).unshared_fields_dims_site()
+        unshared_fields_dims = super().unshared_fields_dims_site()
         unshared_fields_dims.update({
             'agc_status_word': [],
             'lp_status_word': [],
@@ -2092,10 +2036,10 @@ class BorealisRawacf(BorealisRawacfv0_5):
         return unshared_fields_dims
 
 
-class BorealisBfiq(BorealisBfiqv0_5):
+class BorealisBfiqv0_6(BorealisBfiqv0_5):
     """
-    Class containing Borealis Bfiq data fields and their types for the
-    current version of Borealis (v0.6).
+    Class containing Borealis Bfiq data fields and their types for Borealis
+    version 0.6.
 
     See Also
     --------
@@ -2120,47 +2064,7 @@ class BorealisBfiq(BorealisBfiqv0_5):
 
     pulse_phase_offset was also added
     """
-
-    @classmethod
-    def single_element_types(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Returns
-        -------
-        single_element_types
-            All the single-element fields in records of the
-            format, as a dictionary fieldname : type.
-
-        Notes
-        -----
-        In Borealis v0.6, gps_locked, gps_to_system_time_diff, agc_status_word,
-        and lp_status_word were added as single element fields
-        """
-        single_element_types = super(BorealisBfiq,
-                                     cls).single_element_types()
-        single_element_types.update({
-            # the agc fault status of each transmitter, transmitter/USRP
-            # mapped to bit position
-            # A '1' indicates an agc fault at least once during the
-            # integration period.
-            "agc_status_word": np.uint32,
-            # the low power status of each transmitter, transmitter/USRP
-            # mapped to bit position
-            # A '1' indicates a low power condition at least once during
-            # the integration period.
-            "lp_status_word": np.uint32,
-            # Boolean indicating if the GPS was locked during the entire
-            # integration period
-            "gps_locked": np.bool_,
-            # The max time diffe between GPS and system time during the
-            # integration period. In seconds. Negative if GPS time ahead.
-            "gps_to_system_time_diff": np.float64,
-            # Updated to 16 bit number to avoid mismatch when converting
-            # to DMAP format.
-            "experiment_id": np.int16
-        })
-        return single_element_types
+    fields = BorealisFieldsv0_6
 
     @classmethod
     def unshared_fields_dims_array(cls):
@@ -2174,8 +2078,7 @@ class BorealisBfiq(BorealisBfiqv0_5):
         All are unshared fields because their values may not be the same from
         record to record.
         """
-        unshared_fields_dims = super(BorealisBfiq,
-                                     cls).unshared_fields_dims_array()
+        unshared_fields_dims = super().unshared_fields_dims_array()
         unshared_fields_dims.update({
             'agc_status_word': [],
             'lp_status_word': [],
@@ -2190,8 +2093,7 @@ class BorealisBfiq(BorealisBfiqv0_5):
         """
         See BaseFormat class for description and use of this method.
         """
-        unshared_fields_dims = super(BorealisBfiq,
-                                     cls).unshared_fields_dims_site()
+        unshared_fields_dims = super().unshared_fields_dims_site()
         unshared_fields_dims.update({
             'agc_status_word': [],
             'lp_status_word': [],
@@ -2207,10 +2109,10 @@ class BorealisBfiq(BorealisBfiqv0_5):
         return unshared_fields_dims
 
 
-class BorealisAntennasIq(BorealisAntennasIqv0_5):
+class BorealisAntennasIqv0_6(BorealisAntennasIqv0_5):
     """
     Class containing Borealis Antennas iq data fields and their types for
-    Borealis current version (v0.6).
+    Borealis version 0.6.
 
     See Also
     --------
@@ -2235,47 +2137,7 @@ class BorealisAntennasIq(BorealisAntennasIqv0_5):
 
     pulse_phase_offset was also added to the site-structured files.
     """
-
-    @classmethod
-    def single_element_types(cls):
-        """
-        See BaseFormat class for description and use of this method.
-
-        Returns
-        -------
-        single_element_types
-            All the single-element fields in records of the
-            format, as a dictionary fieldname : type.
-
-        Notes
-        -----
-        In Borealis v0.6, gps_locked, gps_to_system_time_diff, agc_status_word,
-        and lp_status_word were added as single element fields
-        """
-        single_element_types = super(BorealisAntennasIq,
-                                     cls).single_element_types()
-        single_element_types.update({
-            # the agc fault status of each transmitter, transmitter/USRP
-            # mapped to bit position
-            # A '1' indicates an agc fault at least once during the
-            # integration period.
-            "agc_status_word": np.uint32,
-            # the low power status of each transmitter, transmitter/USRP
-            # mapped to bit position
-            # A '1' indicates a low power condition at least once during
-            # the integration period.
-            "lp_status_word": np.uint32,
-            # Boolean indicating if the GPS was locked during the entire
-            # integration period
-            "gps_locked": np.bool_,
-            # The max time diffe between GPS and system time during the
-            # integration period. In seconds. Negative if GPS time ahead.
-            "gps_to_system_time_diff": np.float64,
-            # Updated to 16 bit number to avoid mismatch when converting
-            # to DMAP format.
-            "experiment_id": np.int16
-        })
-        return single_element_types
+    fields = BorealisFieldsv0_6
 
     @classmethod
     def unshared_fields_dims_array(cls):
@@ -2289,8 +2151,7 @@ class BorealisAntennasIq(BorealisAntennasIqv0_5):
         All are unshared fields because their values may not be the same from
         record to record.
         """
-        unshared_fields_dims = super(BorealisAntennasIq,
-                                     cls).unshared_fields_dims_array()
+        unshared_fields_dims = super().unshared_fields_dims_array()
         unshared_fields_dims.update({
             'agc_status_word': [],
             'lp_status_word': [],
@@ -2305,8 +2166,7 @@ class BorealisAntennasIq(BorealisAntennasIqv0_5):
         """
         See BaseFormat class for description and use of this method.
         """
-        unshared_fields_dims = super(BorealisAntennasIq,
-                                     cls).unshared_fields_dims_site()
+        unshared_fields_dims = super().unshared_fields_dims_site()
         unshared_fields_dims.update({
             'agc_status_word': [],
             'lp_status_word': [],
@@ -2322,10 +2182,10 @@ class BorealisAntennasIq(BorealisAntennasIqv0_5):
         return unshared_fields_dims
 
 
-class BorealisRawrf(BorealisRawrfv0_5):
+class BorealisRawrfv0_6(BorealisRawrfv0_5):
     """
-    Class containing Borealis Rawrf data fields and their types for current
-    Borealis version (v0.6).
+    Class containing Borealis Rawrf data fields and their types for Borealis 
+    version 0.6.
 
     See Also
     --------
@@ -2342,47 +2202,423 @@ class BorealisRawrf(BorealisRawrfv0_5):
     gps_locked, gps_to_system_time_diff, agc_status_word, and
     lp_status_word.
     """
+    fields = BorealisFieldsv0_6
 
-    @classmethod
-    def single_element_types(cls):
+# The following are the currently used classes, with additions according
+# to Borealis updates.
+
+class BorealisRawacf(BorealisRawacfv0_6):
+    """
+    Class containing Borealis Rawacf data fields and their types for the
+    current version of Borealis (v0.7).
+
+    See Also
+    --------
+    BaseFormat
+    BorealisRawacfv0_6
+    https://borealis.readthedocs.io/en/latest/borealis_data.html
+
+    Notes
+    -----
+    Rawacf data has been mixed, filtered, and decimated; beamformed and
+    combined into antenna arrays; then autocorrelated and correlated between
+    antenna arrays to produce matrices of num_ranges x num_lags.
+
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
+
+    In v0.7, the fields correlation_descriptors and correlation-dimensions
+    were replaced by data_descriptors and data_dimensions, respectively.
+    The field tx_antenna_phases was also added.
+    """
+    fields = BorealisFields
+
+    @staticmethod
+    def find_num_ranges(records: OrderedDict) -> int:
         """
-        See BaseFormat class for description and use of this method.
+        Find the number of ranges given the records dictionary, for
+        restructuring to arrays.
+
+        Parameters
+        ----------
+        records
+            The records dictionary from a site-style file.
 
         Returns
         -------
-        single_element_types
-            All the single-element fields in records of the
-            format, as a dictionary fieldname : type.
+        num_ranges
+            The number of ranges being calculated in the acfs.
 
         Notes
         -----
-        In Borealis v0.6, gps_locked, gps_to_system_time_diff, agc_status_word,
-        and lp_status_word were added as single element fields
+        Num_ranges is unique to a slice so cannot change inside file.
         """
-        single_element_types = super(BorealisRawrf,
-                                     cls).single_element_types()
-        single_element_types.update({
-            # the agc fault status of each transmitter, transmitter/USRP
-            # mapped to bit position
-            # A '1' indicates an agc fault at least once during the
-            # integration period.
-            "agc_status_word": np.uint32,
-            # the low power status of each transmitter, transmitter/USRP
-            # mapped to bit position
-            # A '1' indicates a low power condition at least once during
-            # the integration period.
-            "lp_status_word": np.uint32,
-            # Boolean indicating if the GPS was locked during the entire
-            # integration period
-            "gps_locked": np.bool_,
-            # The max time diffe between GPS and system time during the
-            # integration period. In seconds. Negative if GPS time ahead.
-            "gps_to_system_time_diff": np.float64,
-            # Updated to 16 bit number to avoid mismatch when converting
-            # to DMAP format.
-            "experiment_id": np.int16
-        })
-        return single_element_types
+        first_key = list(records.keys())[0]
+        num_ranges = records[first_key]['data_dimensions'][1]
+        return num_ranges
+
+    @staticmethod
+    def find_num_lags(records: OrderedDict) -> int:
+        """
+        Find the number of lags given the records dictionary, for
+        restructuring to arrays.
+
+        Parameters
+        ----------
+        records
+            The records dictionary from a site-style file.
+
+        Returns
+        -------
+        num_lags
+            The number of lags being calculated in the acfs.
+
+        Notes
+        -----
+        Num_lags is unique to a slice so cannot change inside file.
+        """
+        first_key = list(records.keys())[0]
+        num_lags = records[first_key]['data_dimensions'][2]
+        return num_lags
+
+    @classmethod
+    def site_specific_fields_generate(cls):
+        """
+        See BaseFormat class for description and use of this method.
+        """
+        fields_generate = super().site_specific_fields_generate()
+        data_dimensions = fields_generate.pop('correlation_dimensions')
+        fields_generate['data_dimensions'] = data_dimensions
+        fields_generate.pop('correlation_descriptors')
+        fields_generate['data_descriptors'] = lambda arrays, record_num: np.bytes_(
+            ['num_beams', 'num_ranges', 'num_lags'])
+
+        return fields_generate
+
+    @classmethod
+    def array_specific_fields_generate(cls):
+        """
+        See BaseFormat class for description and use of this method.
+        """
+        fields_generate = super().array_specific_fields_generate()
+        fields_generate.pop('correlation_descriptors')
+        fields_generate['data_descriptors'] = lambda records: np.bytes_(
+            ['num_records', 'max_num_beams', 'num_ranges', 'num_lags'])
+        return fields_generate
+
+    @staticmethod
+    def reshape_site_arrays(records: OrderedDict) -> OrderedDict:
+        """
+        See BaseFormat class for description and use of this method.
+
+        Parameters
+        ----------
+        records
+            An OrderedDict of the site style data, organized
+            by record. Records are stored with timestamps
+            as the keys and the data for that timestamp
+            stored as a dictionary.
+
+        Returns
+        -------
+        records
+            An OrderedDict of the site style data, with the main_acfs,
+            intf_acfs, and xcfs fields in all records reshaped to the correct
+            dimensions.
+
+        Notes
+        -----
+        BorealisRawacf has the correlation fields non-flattened, so nothing needs to be done.
+        """
+        # dimensions provided in data_dimensions field as num_beams,
+        # num_ranges, num_lags for the rawacf format.
+        new_records = copy.deepcopy(records)
+        return new_records
+
+    @staticmethod
+    def flatten_site_arrays(records: OrderedDict) -> OrderedDict:
+        """
+        See BaseFormat class for description and use of this method.
+
+        Parameters
+        ----------
+        records
+            An OrderedDict of the site style data, organized
+            by record. Records are stored with timestamps
+            as the keys and the data for that timestamp
+            stored as a dictionary.
+
+        Returns
+        -------
+        records
+            An OrderedDict of the site style data.
+
+        Notes
+        -----
+        BorealisRawacf has the main_acfs, intf_acfs, and xcfs fields non-flattened
+        in the site structured files, so nothing needs to be done.
+        """
+        new_records = copy.deepcopy(records)
+        return new_records
+
+    @classmethod
+    def site_get_max_dims(cls, filename: str, unshared_parameters: List[str]):
+        """
+        See BaseFormat class for description and use of this method.
+
+        Parameters
+        ----------
+        filename: str
+            Name of the site file being checked
+        unshared_parameters: List[str]
+            List of parameter names that are not shared between all the records
+            in the site restructured file, i.e. may have different dimensions
+            between records.
+
+        Returns
+        -------
+        fields_max_dims: dict
+            dictionary containing field names (str) as keys with maximum
+            dimensions required to restructure to array file as values (tuples)
+        """
+        fields_max_dims, max_num_sequences, max_num_beams = BaseFormat.site_get_max_dims(filename, unshared_parameters)
+
+        # Now change the main_acfs, int_acfs and xcfs dicts to maximum required dims
+
+        # Get the num_ranges and num_lags fields directly from one record of the file
+        with h5py.File(filename, 'r') as site_file:
+            # hacky way to get first key with KeyView object from .keys()
+            record_name = [k for i, k in enumerate(site_file.keys()) if i == 0][0]
+            _, num_ranges, num_lags = site_file[record_name]['data_dimensions']
+
+        # Change the data dimensions to the multidimensional size instead of flattened size
+        reshaped_correlation_dims = (max_num_beams, num_ranges, num_lags)
+        fields_max_dims['main_acfs'] = reshaped_correlation_dims
+        fields_max_dims['intf_acfs'] = reshaped_correlation_dims
+        fields_max_dims['xcfs'] = reshaped_correlation_dims
+
+        return fields_max_dims, max_num_sequences, max_num_beams
+
+    @classmethod
+    def unshared_fields_dims_array(cls):
+        """
+        See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        In Borealis v0.7, tx_antenna_phases was added to rawacf. It is an
+        unshared field because its values may not be the same from record to record.
+        """
+        unshared_fields_dims = super().unshared_fields_dims_array()
+        unshared_fields_dims.update({
+            'tx_antenna_phases': [cls.fields.find_num_antennas_site],
+            })
+        return unshared_fields_dims
+
+    @classmethod
+    def unshared_fields_dims_site(cls):
+        """
+        See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        In Borealis v0.7, tx_antenna_phases was added to rawacf. It is an
+        unshared field because its values may not be the same from record to record.
+        """
+        unshared_fields_dims = super().unshared_fields_dims_site()
+        unshared_fields_dims.update({
+            'tx_antenna_phases': [lambda arrays, record_num: arrays['tx_antenna_phases'].shape[1]],
+            })
+        return unshared_fields_dims
+
+
+class BorealisBfiq(BorealisBfiqv0_6):
+    """
+    Class containing Borealis Bfiq data fields and their types for the
+    current version of Borealis (v0.7).
+
+    See Also
+    --------
+    BaseFormat
+    BorealisBfiqv0_6
+
+    Notes
+    -----
+    Bfiq data is beamformed i and q data. It has been mixed, filtered,
+    decimated to the final output receive rate, and it has been beamformed
+    and all channels have been combined into their arrays. No correlation
+    or averaging has occurred.
+
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
+
+    The field tx_antenna_phases was added to the bfiq format in v0.7.
+    """
+    fields = BorealisFields
+
+    @classmethod
+    def unshared_fields_dims_array(cls):
+        """
+        See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        In Borealis v0.7, tx_antenna_phases was added to bfiq. It is an
+        unshared field because its values may not be the same from record to record.
+        """
+        unshared_fields_dims = super().unshared_fields_dims_array()
+        unshared_fields_dims.update({
+            'tx_antenna_phases': [cls.fields.find_num_antennas_site],
+            })
+        return unshared_fields_dims
+
+    @classmethod
+    def unshared_fields_dims_site(cls):
+        """
+        See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        In Borealis v0.7, tx_antenna_phases was added to bfiq. It is an
+        unshared field because its values may not be the same from record to record.
+        """
+        unshared_fields_dims = super().unshared_fields_dims_site()
+        unshared_fields_dims.update({
+            'tx_antenna_phases': [lambda arrays, record_num: arrays['tx_antenna_phases'].shape[1]],
+            })
+        return unshared_fields_dims
+
+
+class BorealisAntennasIq(BorealisAntennasIqv0_6):
+    """
+    Class containing Borealis Antennas iq data fields and their types for
+    Borealis current version (v0.7).
+
+    See Also
+    --------
+    BaseFormat
+    BorealisAntennasIqv0_6
+
+    Notes
+    -----
+    Antennas iq data is data with all channels separated. It has been mixed
+    and filtered, but it has not been beamformed or combined into the
+    entire antenna array data product.
+
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
+
+    In v0.7, the following fields were added to the Borealis-produced
+    site structured files for ease of postprocessing:
+    first_range
+    first_range_rtt
+    lags
+    num_ranges
+    range_sep
+
+    The field tx_antenna_phases was added to the antennas_iq format in v0.7.
+    """
+    fields = BorealisFields
+
+    @classmethod
+    def unshared_fields_dims_array(cls):
+        """
+        See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        In Borealis v0.7, tx_antenna_phases was added to antennas_iq. It is an
+        unshared field because its values may not be the same from record to record.
+        """
+        unshared_fields_dims = super().unshared_fields_dims_array()
+        unshared_fields_dims.update({
+            'tx_antenna_phases': [cls.find_num_antennas],
+            })
+        return unshared_fields_dims
+
+    @classmethod
+    def unshared_fields_dims_site(cls):
+        """
+        See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        In Borealis v0.7, tx_antenna_phases was added to antennas_iq. It is an
+        unshared field because its values may not be the same from record to record.
+        """
+        unshared_fields_dims = super().unshared_fields_dims_site()
+        unshared_fields_dims.update({
+            'tx_antenna_phases': [lambda arrays, record_num: arrays['data'].shape[1]],
+            })
+        return unshared_fields_dims
+
+    @classmethod
+    def site_specific_fields_generate(cls):
+        """
+        See BaseFormat class for description and use of this method.
+        """
+        return {
+            'data_descriptors': lambda arrays, record_num: np.bytes_(
+                ['num_antennas', 'num_sequences', 'num_samps']),
+            'data_dimensions': lambda arrays, record_num: np.array(
+                [arrays['data'].shape[1], arrays['num_sequences'][record_num],
+                 arrays['data'].shape[3]], dtype=np.uint32)
+        }
+
+class BorealisRawrf(BorealisRawrfv0_6):
+    """
+    Class containing Borealis Rawrf data fields and their types for current
+    Borealis version (v0.7).
+
+    See Also
+    --------
+    BaseFormat
+    BorealisRawrfv0_6
+
+    Notes
+    -----
+    See BaseFormat for description of classmethods and how they
+    are used to verify format files and restructure Borealis files to
+    array and site structure.
+
+    The field tx_antenna_phases was added to the rawrf format in v0.7.
+    """
+    fields = BorealisFields
+
+    @classmethod
+    def unshared_fields_dims_array(cls):
+        """
+        See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        In Borealis v0.7, tx_antenna_phases was added to rawrf. It is an
+        unshared field because their values may not be the same from record to record.
+        """
+        unshared_fields_dims = super().unshared_fields_dims_array()
+        unshared_fields_dims.update({
+            'tx_antenna_phases': [cls.find_num_antennas],
+            })
+        return unshared_fields_dims
+
+    @classmethod
+    def unshared_fields_dims_site(cls):
+        """
+        See BaseFormat class for description and use of this method.
+
+        Notes
+        -----
+        In Borealis v0.7, tx_antenna_phases was added to rawrf. It is an
+        unshared field because their values may not be the same from record to record.
+        """
+        unshared_fields_dims = super().unshared_fields_dims_site()
+        unshared_fields_dims.update({
+            'tx_antenna_phases': [lambda arrays, record_num: arrays['data'].shape[1]],
+            })
+        return unshared_fields_dims
 
 
 # borealis versions
@@ -2412,6 +2648,12 @@ borealis_version_dict = {
         'rawrf': BorealisRawrfv0_5
         },
     'v0.6': {
+        'bfiq': BorealisBfiqv0_6,
+        'rawacf': BorealisRawacfv0_6,
+        'antennas_iq': BorealisAntennasIqv0_6,
+        'rawrf': BorealisRawrfv0_6
+        },
+    'v0.7': {
         'bfiq': BorealisBfiq,
         'rawacf': BorealisRawacf,
         'antennas_iq': BorealisAntennasIq,
